# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SaaS Starter API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  clerkUserId: string;
  email: string;
  createdAt: string;
}

export interface UpsertUserBody {
  email: string;
}